requirements.txt
fastapi
uvicorn
openai
faiss-cpu
python-dotenv
tqdm
